Refactor validation logic into helper
